add IssueCard component
This breaks a few things but i'm going
to roll with it for a bit to see where this leads.

I need to rethink the way I structured the
columns. Maybe I could use lodash to group them?

I plan to refactor the IssueCard component into
smaller pieces, but it helped to refactor using a
similar structure as the original view.