add lodash-rails gem
Working with jQueryUi and React is going to 
require some planning. It is possible to do from
some blog posts i've read, and i'm not sure my
initial thoughts will work.

I am planning to use Lodash to group the issues by
column to display the initial state. Then I need
to check the docs for jQueryUi and React to see
what they have to offer.