change data structure in repos#show controller
I'm experimenting with passing a repo object that
contains the information for the repo and issues.

I think this might help to remove some clutter in
the code. I am also experimenting with OpenStruct
to pass just what I want from the github api.

I need to understand how this table method works
as well. I'm not sure if I like this result but
I am going to proceed for a bit to see where it
goes.