remove data and id's from html
I have been trying to refactor and preserve the
way the app looks. This is making it more
difficult, and since my goal is to remove all
of the jQuery and use React it is pointless to
leave it. It will be easier to look at this
without the distraction, and I can add back what
is needed later for styling.